Reverse the output from _follow_also
Turns out when we follow the export tree we return the list in the order we
discovered, which is the inverse of the order we want. This wasn't ever a
problem before because nobody has tried exporting more than one level off of
Moose. Someone has now, so we need to reverse the entire list.